fixed agile tests
